+++ src/main/java/edu/ksu/canvas/interfaces/MigrationIssueReader.java
@@ -15,7 +15,7 @@ public interface MigrationIssueReader extends CanvasReader<MigrationIssue, Migra
      * @return List of migration issues
      * @throws IOException When there is an error communicating with Canvas
      */
-    List<MigrationIssue> getCourseMigrationIssues(String courseId, Integer migrationId) throws IOException;
+    List<MigrationIssue> getCourseMigrationIssues(String courseId, Long migrationId) throws IOException;
 
     /**
      * Returns a migration issue.
@@ -25,7 +25,7 @@ public interface MigrationIssueReader extends CanvasReader<MigrationIssue, Migra
      * @return A migration issue
      * @throws IOException When there is an error communicating with Canvas
      */
-    Optional<MigrationIssue> getCourseMigrationIssue(String courseId, Integer migrationId, Integer issueId) throws IOException;
+    Optional<MigrationIssue> getCourseMigrationIssue(String courseId, Long migrationId, Long issueId) throws IOException;
 
     /**
      * Returns a list migration issues.
@@ -34,7 +34,7 @@ public interface MigrationIssueReader extends CanvasReader<MigrationIssue, Migra
      * @return List of migration issues
      * @throws IOException When there is an error communicating with Canvas
      */
-    List<MigrationIssue> getUserMigrationIssues(String userId, Integer migrationId) throws IOException;
+    List<MigrationIssue> getUserMigrationIssues(String userId, Long migrationId) throws IOException;
 
     /**
      * Returns a migration issue.
@@ -44,7 +44,7 @@ public interface MigrationIssueReader extends CanvasReader<MigrationIssue, Migra
      * @return A migration issue
      * @throws IOException When there is an error communicating with Canvas
      */
-    Optional<MigrationIssue> getUserMigrationIssue(String userId, Integer migrationId, Integer issueId) throws IOException;
+    Optional<MigrationIssue> getUserMigrationIssue(String userId, Long migrationId, Long issueId) throws IOException;
 
     /**
      * Returns a list migration issues.
@@ -53,7 +53,7 @@ public interface MigrationIssueReader extends CanvasReader<MigrationIssue, Migra
      * @return List of migration issues
      * @throws IOException When there is an error communicating with Canvas
      */
-    List<MigrationIssue> getGroupMigrationIssues(String groupId, Integer migrationId) throws IOException;
+    List<MigrationIssue> getGroupMigrationIssues(String groupId, Long migrationId) throws IOException;
 
     /**
      * Returns a migration issue.
@@ -63,7 +63,7 @@ public interface MigrationIssueReader extends CanvasReader<MigrationIssue, Migra
      * @return A migration issue
      * @throws IOException When there is an error communicating with Canvas
      */
-    Optional<MigrationIssue> getGroupMigrationIssue(String groupId, Integer migrationId, Integer issueId) throws IOException;
+    Optional<MigrationIssue> getGroupMigrationIssue(String groupId, Long migrationId, Long issueId) throws IOException;
 
     /**
      * Returns a list migration issues.
@@ -72,7 +72,7 @@ public interface MigrationIssueReader extends CanvasReader<MigrationIssue, Migra
      * @return List of migration issues
      * @throws IOException When there is an error communicating with Canvas
      */
-    List<MigrationIssue> getAccountMigrationIssues(String accountId, Integer migrationId) throws IOException;
+    List<MigrationIssue> getAccountMigrationIssues(String accountId, Long migrationId) throws IOException;
 
     /**
      * Returns a migration issue.
@@ -82,5 +82,5 @@ public interface MigrationIssueReader extends CanvasReader<MigrationIssue, Migra
      * @return A migration issue
      * @throws IOException When there is an error communicating with Canvas
      */
-    Optional<MigrationIssue> getAccountMigrationIssue(String accountId, Integer migrationId, Integer issueId) throws IOException;
+    Optional<MigrationIssue> getAccountMigrationIssue(String accountId, Long migrationId, Long issueId) throws IOException;
 }
