+++ src/main/java/edu/ksu/canvas/model/Course.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class Course extends BaseCanvasModel implements Serializable {
     public static final long serialVersionUID = 1L;
 
-    private Integer accountId;
+    private Long accountId;
     private String courseCode;
     private String defaultView;
     private Long id;
@@ -31,7 +31,7 @@ public class Course extends BaseCanvasModel implements Serializable {
     private String sisCourseId;
     private String integrationId;
     private String workflowState;
-    private Integer totalStudents;
+    private Long totalStudents;
     private Long enrollmentTermId;
     private Boolean restrictEnrollmentsToCourseDates;
     private Long gradingStandardId;
@@ -58,7 +58,7 @@ public class Course extends BaseCanvasModel implements Serializable {
     // Only used when asking for the course image
     private String imageDownloadUrl;
     // Only used when setting course image
-    private Integer imageId;
+    private Long imageId;
     private String imageUrl;
     private String publicDescription;
     // Only used when pushing a course status change to Canvas
@@ -84,11 +84,11 @@ public class Course extends BaseCanvasModel implements Serializable {
     }
 
     @CanvasField(postKey = "account_id")
-    public Integer getAccountId() {
+    public Long getAccountId() {
         return accountId;
     }
 
-    public void setAccountId(Integer accountId) {
+    public void setAccountId(Long accountId) {
         this.accountId = accountId;
     }
 
@@ -213,11 +213,11 @@ public class Course extends BaseCanvasModel implements Serializable {
         this.workflowState = workflowState;
     }
 
-    public Integer getTotalStudents() {
+    public Long getTotalStudents() {
         return totalStudents;
     }
 
-    public void setTotalStudents(Integer totalStudents) {
+    public void setTotalStudents(Long totalStudents) {
         this.totalStudents = totalStudents;
     }
 
@@ -433,11 +433,11 @@ public class Course extends BaseCanvasModel implements Serializable {
     }
 
     @CanvasField(postKey = "image_id")
-    public Integer getImageId() {
+    public Long getImageId() {
         return imageId;
     }
 
-    public void setImageId(Integer imageId) {
+    public void setImageId(Long imageId) {
         this.imageId = imageId;
     }
